requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
dnspython>=2.4.0
urllib3>=2.0.0
